bug9649 - add red test to integration test covering bug expected behavior, extract logic from journey file that is clerk agnostic
